+++ examples/java/demo-layer-url/jsrc/main/java/com/aws/lambda/response/HelloMessage.java
@@ -1,8 +1,0 @@
-package com.aws.lambda.response;
-
-public class HelloMessage {
-    private final String message;
-    public HelloMessage(String message) {
-        this.message = message;
-    }
-}